+++ examples/preact-ts/src/pages/number-input/basic.tsx
@@ -2,12 +2,12 @@ import * as numberInput from "@zag-js/number-input"
 import { normalizeProps, useMachine } from "@zag-js/preact"
 import { numberInputControls } from "@zag-js/shared"
 import { useId } from "react"
-import { useControls } from "../../hooks/use-controls"
-import { Toolbar } from "../../components/toolbar"
 import { StateVisualizer } from "../../components/state-visualizer"
+import { Toolbar } from "../../components/toolbar"
+import { useControls } from "../../hooks/use-controls"
 import "@styles/number-input.css"
 
-export default function NumberInput() {
+export default function Page() {
   const controls = useControls(numberInputControls)
 
   const service = useMachine(numberInput.machine, {
@@ -35,7 +35,6 @@ export default function NumberInput() {
           </div>
         </div>
       </main>
-
       <Toolbar controls={controls.ui}>
         <StateVisualizer state={service} omit={["formatter", "parser"]} />
       </Toolbar>