+++ packages/machines/date-picker/src/date-picker.machine.ts
@@ -362,7 +362,7 @@ export const machine = createMachine<DatePickerSchema>({
       on: {
         "CONTROLLED.OPEN": {
           target: "open",
-          actions: ["focusFirstSelectedDate", "focusActiveCell"],
+          actions: ["resetView", "focusFirstSelectedDate", "focusActiveCell"],
         },
         "TRIGGER.CLICK": [
           {
@@ -371,7 +371,7 @@ export const machine = createMachine<DatePickerSchema>({
           },
           {
             target: "open",
-            actions: ["focusFirstSelectedDate", "focusActiveCell", "invokeOnOpen"],
+            actions: ["resetView", "focusFirstSelectedDate", "focusActiveCell", "invokeOnOpen"],
           },
         ],
         OPEN: [
@@ -381,7 +381,7 @@ export const machine = createMachine<DatePickerSchema>({
           },
           {
             target: "open",
-            actions: ["focusFirstSelectedDate", "focusActiveCell", "invokeOnOpen"],
+            actions: ["resetView", "focusFirstSelectedDate", "focusActiveCell", "invokeOnOpen"],
           },
         ],
       },
@@ -392,7 +392,7 @@ export const machine = createMachine<DatePickerSchema>({
       on: {
         "CONTROLLED.OPEN": {
           target: "open",
-          actions: ["focusFirstSelectedDate", "focusActiveCell"],
+          actions: ["resetView", "focusFirstSelectedDate", "focusActiveCell"],
         },
         "TRIGGER.CLICK": [
           {
@@ -401,7 +401,7 @@ export const machine = createMachine<DatePickerSchema>({
           },
           {
             target: "open",
-            actions: ["focusFirstSelectedDate", "focusActiveCell", "invokeOnOpen"],
+            actions: ["resetView", "focusFirstSelectedDate", "focusActiveCell", "invokeOnOpen"],
           },
         ],
         OPEN: [
@@ -411,7 +411,7 @@ export const machine = createMachine<DatePickerSchema>({
           },
           {
             target: "open",
-            actions: ["focusFirstSelectedDate", "focusActiveCell", "invokeOnOpen"],
+            actions: ["resetView", "focusFirstSelectedDate", "focusActiveCell", "invokeOnOpen"],
           },
         ],
       },
@@ -420,7 +420,7 @@ export const machine = createMachine<DatePickerSchema>({
     open: {
       tags: ["open"],
       effects: ["trackDismissableElement", "trackPositioning"],
-      exit: ["clearHoveredDate", "resetView"],
+      exit: ["clearHoveredDate"],
       on: {
         "CONTROLLED.CLOSE": [
           {
@@ -1203,13 +1203,14 @@ export const machine = createMachine<DatePickerSchema>({
         date = constrainValue(date, prop("min"), prop("max"))
 
         const values = Array.from(context.get("value"))
-        values[event.index] = date
+        values[event.index] = preserveTime(values[event.index], date)
+        const adjustedValues = adjustStartAndEndDate(values)
 
-        context.set("value", values)
+        context.set("value", adjustedValues)
 
         // always sync the input value, even if the selecteddate is not changed
         // e.g. selected value is 02/28/2024, and the input value changed to 02/28
-        const valueAsString = getValueAsString(values, prop)
+        const valueAsString = getValueAsString(adjustedValues, prop)
         context.set("inputValue", valueAsString[event.index])
       },