+++ playwright.config.ts
@@ -13,6 +13,7 @@ export function getWebServer(): WebServer {
   const vuePort = process.env.PORT ?? "3001"
   const solidPort = process.env.PORT ?? "3002"
   const sveltePort = process.env.PORT ?? "3003"
+  const preactPort = process.env.PORT ?? "3004"
 
   const frameworks: Record<string, WebServer> = {
     react: {
@@ -37,6 +38,12 @@ export function getWebServer(): WebServer {
       cwd: "./examples/svelte-ts",
       command: `pnpm vite --port ${sveltePort}`,
       url: `http://localhost:${sveltePort}`,
+      reuseExistingServer: !CI,
+    },
+    preact: {
+      cwd: "./examples/preact-ts",
+      command: `pnpm vite --port ${preactPort}`,
+      url: `http://localhost:${preactPort}`,
       reuseExistingServer: !CI,
     },
   }