+++ packages/machines/accordion/CHANGELOG.md
@@ -21,6 +21,18 @@
   ```ts
   accordion.machine({ loopFocus: false }) // stop at the ends instead of wrapping
   ```
+## 1.42.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/anatomy@1.42.0
+  - @zag-js/core@1.42.0
+  - @zag-js/types@1.42.0
+  - @zag-js/utils@1.42.0
+  - @zag-js/dom-query@1.42.0
+
+## 1.41.2
 
 ### Patch Changes