+++ packages/machines/async-list/CHANGELOG.md
@@ -110,6 +110,15 @@
   ```ts
   async load({ items, cursor, signal }) { ... }
   ```
+## 1.42.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/core@1.42.0
+  - @zag-js/utils@1.42.0
+
+## 1.41.2
 
 ### Patch Changes