+++ packages/machines/avatar/CHANGELOG.md
@@ -7,6 +7,18 @@
 - [`772046f`](https://github.com/chakra-ui/zag/commit/772046ffe214037838d4e55f306286fdc62041d7) Thanks
   [@segunadebayo](https://github.com/segunadebayo)! - Expose the image loading status. Adds `api.status`
   (`loading | loaded | error`) and emits `loading` from `onStatusChange`, so you can render a distinct loading state.
+## 1.42.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/anatomy@1.42.0
+  - @zag-js/core@1.42.0
+  - @zag-js/types@1.42.0
+  - @zag-js/utils@1.42.0
+  - @zag-js/dom-query@1.42.0
+
+## 1.41.2
 
 ### Patch Changes