+++ packages/machines/carousel/CHANGELOG.md
@@ -34,6 +34,19 @@
   - padding: "16px"
   + itemSpacing: "16px"
   ```
+## 1.42.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/anatomy@1.42.0
+  - @zag-js/core@1.42.0
+  - @zag-js/types@1.42.0
+  - @zag-js/utils@1.42.0
+  - @zag-js/dom-query@1.42.0
+  - @zag-js/scroll-snap@1.42.0
+
+## 1.41.2
 
 ### Patch Changes