+++ packages/machines/password-input/CHANGELOG.md
@@ -1,6 +1,18 @@
 # @zag-js/password-input
 
 ## 2.0.0-next.0
+## 1.42.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/anatomy@1.42.0
+  - @zag-js/core@1.42.0
+  - @zag-js/types@1.42.0
+  - @zag-js/utils@1.42.0
+  - @zag-js/dom-query@1.42.0
+
+## 1.41.2
 
 ### Patch Changes