+++ packages/machines/presence/CHANGELOG.md
@@ -1,6 +1,16 @@
 # @zag-js/presence
 
 ## 2.0.0-next.0
+## 1.42.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/core@1.42.0
+  - @zag-js/types@1.42.0
+  - @zag-js/dom-query@1.42.0
+
+## 1.41.2
 
 ### Patch Changes