+++ packages/types/CHANGELOG.md
@@ -1,6 +1,9 @@
 # @zag-js/types
 
 ## 2.0.0-next.0
+## 1.42.0
+
+## 1.41.2
 
 ### Major Changes