+++ packages/utilities/file-utils/CHANGELOG.md
@@ -1,6 +1,14 @@
 # @zag-js/file-utils
 
 ## 2.0.0-next.0
+## 1.42.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/i18n-utils@1.42.0
+
+## 1.41.2
 
 ### Patch Changes