+++ packages/utilities/popper/CHANGELOG.md
@@ -1,6 +1,15 @@
 # @zag-js/popper
 
 ## 2.0.0-next.0
+## 1.42.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/utils@1.42.0
+  - @zag-js/dom-query@1.42.0
+
+## 1.41.2
 
 ### Patch Changes